Made numerous changes and fixes. The limiter "service" now runs as an invisible standalone process (not a service), and the console interface serves only to construct the configuration file for it (in debug mode, the console also launches and controls a simulated process limiter). Also, groups can now select processes based on executable file path rather than just name.

+++ ProcessLimiterSrvc/Group.cs
@@ -12,20 +12,33 @@ using System.Windows.Forms;
 using System.Drawing;
 using System.Reflection;
 
+using System.Speech.Synthesis;
+
 namespace ProcessLimiterSrvc
 {
     [Serializable]
     public class Group
     {
+        [NonSerialized]
+        internal static SpeechSynthesizer Speaker = new SpeechSynthesizer();
+
         public Group(string Name)
         { this.Name = Name; }
 
         public readonly List<string> ProcessNames = new List<string>();
+        public readonly List<string> DirectoryNames = new List<string>(); 
         public TimeSpan? TimeLimit = null;
         public TimeSpan? StartTime = null, EndTime = null;
         public readonly string Name = "";
 
+        [NonSerialized]
+        private TimeSpan? TimeRemaining = null;
+        [NonSerialized]
+        private static TimeSpan? WarningBoundary = TimeSpan.FromMinutes(15);
+
+        [NonSerialized]
         public DateTime LastUpdate;
+        [NonSerialized]
         private TimeSpan? LimitRemaining = null, BoundaryRemaining = null;
 
         private static Regex MatchWildcard(string SimplePattern)
@@ -36,11 +49,43 @@ namespace ProcessLimiterSrvc
                        + "$", RegexOptions.IgnoreCase);
         }
 
+        private static Regex DirectoryWildcard(string SimplePattern)
+        {
+            return new Regex("^" + Regex.Escape(SimplePattern)
+                              .Replace(@"\*", ".*")
+                              .Replace(@"\?", "."), RegexOptions.IgnoreCase);
+        }
+
         public void Initialize()
         {
             LastUpdate = DateTime.Now;
             LimitRemaining = TimeLimit;
             Update();
+        }
+
+        private bool MatchName(Process p)
+        {
+            return ProcessNames
+                .Select(n => MatchWildcard(n))
+                .Any(r => r.IsMatch(p.ProcessName));
+        }
+
+        private bool MatchDirectory(Process p)
+        {
+            return DirectoryNames
+                .Select(n => DirectoryWildcard(n))
+                .Any(r =>
+                {
+                    try
+                    {
+                        return r.IsMatch(p.MainModule.FileName);
+                    }
+
+                    catch(Exception ex) when (ex is Win32Exception || ex is InvalidOperationException)
+                    {
+                        return false;
+                    }
+                });
         }
 
         public Process[] RunningProcs
@@ -49,10 +94,7 @@ namespace ProcessLimiterSrvc
             {
                 return
                     Process.GetProcesses().Where(p =>
-                        ProcessNames.Select(n =>
-                            MatchWildcard(n))
-                            .Any(r =>
-                                r.IsMatch(p.ProcessName))).ToArray();
+                        MatchName(p) || MatchDirectory(p)).ToArray();
             }
         }
 
@@ -85,6 +127,7 @@ namespace ProcessLimiterSrvc
             bool IsNewDay = !DateTime.Now.Date.Equals(LastUpdate.Date);
             var Running = RunningProcs;
             bool NeedToTerminate = false;
+            TimeSpan? NewTimeRemaining = null;
 
             if (Running.Length > 0)
             {
@@ -96,7 +139,7 @@ namespace ProcessLimiterSrvc
                         BoundaryRemaining = TimeSpan.Zero;
                 else if (StartTime != null)
                     BoundaryRemaining = NewTime > StartTime ? TimeSpan.FromDays(1) - NewTime : TimeSpan.Zero;
-                else if (EndTime != null && EndTime < NewTime)
+                else if (EndTime != null)
                     BoundaryRemaining = EndTime - NewTime;
 
                 if (TimeLimit != null)
@@ -109,29 +152,60 @@ namespace ProcessLimiterSrvc
                         LimitRemaining = TimeLimit;
                 }
 
-                if (LimitRemaining != null && BoundaryRemaining != null)
+                if (LimitRemaining.HasValue && BoundaryRemaining.HasValue)
                 {
-                    NeedToTerminate = !(LimitRemaining > TimeSpan.Zero && BoundaryRemaining > TimeSpan.Zero);
+                    NewTimeRemaining = LimitRemaining < BoundaryRemaining ? LimitRemaining : BoundaryRemaining;
                 }
-                else if (LimitRemaining != null)
+                else if (LimitRemaining.HasValue)
                 {
-                    NeedToTerminate = !(LimitRemaining > TimeSpan.Zero);
+                    NewTimeRemaining = LimitRemaining;
                 }
-                else if (BoundaryRemaining != null)
+                else if (BoundaryRemaining.HasValue)
                 {
-                    NeedToTerminate = !(BoundaryRemaining > TimeSpan.Zero);
+                    NewTimeRemaining = BoundaryRemaining;
                 }
+                NeedToTerminate = !(NewTimeRemaining > TimeSpan.Zero);
+                
+                SpeakWarning(TimeRemaining, NewTimeRemaining, Running);
 
                 if (NeedToTerminate)
                 {
                     KillAll(Running);
                 }
             }
+            TimeRemaining = NewTimeRemaining;
             LastUpdate = DateTime.Now;
+        }
+
+        private void SpeakWarning(TimeSpan? timeRemaining, TimeSpan? newTimeRemaining, Process[] procs)
+        {
+            var ProcNames = procs.Select(p => p.ProcessName.Replace(".", " dot ")).ToArray();
+            string SayProcs;
+            if (ProcNames.Length == 0)
+                return;
+            else if (ProcNames.Length == 1)
+                SayProcs = ProcNames[0];
+            else if (ProcNames.Length == 2)
+                SayProcs = string.Format("{0} and {1}", ProcNames);
+            else
+                SayProcs = string.Format("{0} processes", ProcNames.Length);
+
+            string SayTime = "unknown time";
+            if (newTimeRemaining.HasValue)
+                if (newTimeRemaining > TimeSpan.Zero)
+                    SayTime = string.Format("{0} minutes", newTimeRemaining.Value.Minutes);
+
+            if (newTimeRemaining.HasValue)
+                if (newTimeRemaining <= TimeSpan.Zero)
+                    Speaker.SpeakAsync($"Closing {SayProcs}");
+                else if ((!timeRemaining.HasValue || timeRemaining > WarningBoundary) &&
+                         WarningBoundary >= newTimeRemaining)
+                    Speaker.SpeakAsync($"{SayProcs} will close in {SayTime}");
         }
 
         private void KillAll(IEnumerable<Process> Procs)
         {
+            var Killed = new List<string>();
             foreach (var proc in Procs)
             {
                 Task t = new Task(() =>
@@ -140,24 +214,36 @@ namespace ProcessLimiterSrvc
                     try
                     {
                         proc.Kill();
-                        note.BalloonTipTitle = "ProcessLimiter Notice";
-                        note.BalloonTipText = "ProcessLimiter has terminated '" + proc.ProcessName + "'";
+                        Killed.Add(proc.ProcessName);
                     }
                     catch (Win32Exception)
                     {
-                        note.BalloonTipTitle = "ProcessLimiter Permission Denied";
-                        note.BalloonTipText = "ProcessLimiter denied permission to kill process '" + proc.ProcessName + "'";
+                        if (!proc.HasExited)
+                        {
+                            note.BalloonTipTitle = "ProcessLimiter Permission Denied";
+                            note.BalloonTipText = "ProcessLimiter denied permission to kill process '" + proc.ProcessName + "'";
+                        }
                     }
                     catch (InvalidOperationException)
                     { /* Ignore */ return; }
                     note.Text = "ProcessLimiter";
                     note.Icon = new Icon(Assembly.GetExecutingAssembly().GetManifestResourceStream("ProcessLimiterSrvc.clock.ico"));
                     note.Visible = true;
-                    note.ShowBalloonTip(3000);
+                    note.ShowBalloonTip(1000);
                     Application.Run();
                 });
                 t.Start();
             }
+            var killedNote = new NotifyIcon();
+            killedNote.BalloonTipTitle = "ProcessLimiter Notice";
+            killedNote.BalloonTipText = "ProcessLimiter has terminated the following:";
+            foreach (var name in Killed)
+                killedNote.BalloonTipText += $" '{name}'";
+            killedNote.Text = "ProcessLimiter";
+            killedNote.Icon = new Icon(Assembly.GetExecutingAssembly().GetManifestResourceStream("ProcessLimiterSrvc.clock.ico"));
+            killedNote.Visible = true;
+            killedNote.ShowBalloonTip(1000);
+            Application.Run();
         }
 
         public void KillAll()
